Prettied up the code even more.
Originally, it wasn't perfectly in line with the style guidelines posted by prof, so I fixed that. Also, it turns out that the way VSC shows you code isn't exactly how it looks in the raw file. When you guys download this, the indentations of the asterisks for those block codes are probably gonna be all messed up. However, if you look at the code here on Github (and if you opened it in notepad or a text editor like Vim) it will would look just fine. The take away here is that Microsoft sucks