Completed Classes !PLEASE READ DESCRIPTION!
Fully implemented the mastermind and response classes.

I have defined all required functions for the mastermind class, and implemented it in a main function which can be found in test.cpp, and run with the a.out file. I also had to edit the response class, and I overloaded the constructor for it (very likely unnecessary, especially since I only use 1 version of the constructor, and never the overloaded version). I also edited the get and set functions. IMPORTANT NOTE: you cannot return multiple values with a single return statement! ``return (x, y);`` will not work (and should give you an error, idk why it didn't here). We should consider getting rid of the set) and get() functions, and stick with the setting and getting of the individual values. Currently, the game is completely playable and in a submittable stage, all that is left to do is to move the contents of test.cpp to a main file, and add comments as well as to check for formatting.