Uploaded Deck Class and main
The deck class is completed with everything but the shuffle function. I have also implemented it in a main function. You can run the test file in the terminal to verify. Do note that the deck class relies on the card header file to compile, so if you want to compile it yourself, you'll need to also obtain card.h