Add files via upload
Current version of the shuffle function randomly picks a pair and shuffles the two nodes within that "pair".  It randomly picks a number between 2 and 51, traverses the linked list to that location, and shuffles the pair at that location. Major thing to note is that this algorithm doesn't include the shuffling of the head or last node. That functionality needs to be added.

Note: I am considering upgrading the shuffle algorithm to randomly pick two nodes. Then shuffling those two chosen nodes. I believe this will lead to better shuffling. 

My version of the deck and card classes are there as well. Major difference there is the implemention of the deck class.