Created Replace Function
I have implemented a working replace function. I also fixed the `<<` operator, and I changed the constructor to be able to take integer inputs such that we can make decks of different sizes as necessary. Do note, that if you make a deck with 0 cards, and try to replace a card, you get a seg fault because, you are trying to set the new next value of front, when front is equal to NULL.