Updated `<<` Operator
I made the `<<` Operator even more modular. Before, it relied on the deck being exactly 52 cards. But now, it just prints out cards as long as the temporary list found within the function is not equal to NULL. It still prints in rows of 13. Also, the constructor still takes an `int` parameter to determine how many cards are in a deck when initialized. We could potentially just overload the constructor with a version that can just make an "empty deck," or, we can just make a function that forces a deck to be empty.