Created Fully Working Deck Class
Here I have completely implemented Rohan's deal function (I modified it slightly, but the actual algorithm and method is still exactly the same), as well as the replace function. With this deck class, we can make a deck with as many or as few cards as we want (though, no fewer than zero cards may be in a deck), and then add cards into that deck (regardless of size) with replace, and then print out the whole deck using the `<<` operator (again, regardless of deck size). Then, for decks that have more than 0 cards in them, the deal will 'pop' the last card and return its value.