Update dictionary.h
The dictionary class should be done! Please note that each word in the `words` vector has `\n` at the end. For example, if you want "banana" inside `words`, the code will have it as `"banana\n"`. The reason for this is explained in the `readDict()` function, but it's pretty simple. The dictionary we get has every word on its own line, so I wanted to mimic that here. So, since the end line character is appended to every string, whenever we print a string from `words`, it will automatically end line.